% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_BFR.R
\docType{data}
\name{intercalibration_BFR}
\alias{intercalibration_BFR}
\title{Intercalibration data for BFR (BDE28, BDE47, BDE99, BDE100, BDE153, BDE154,  HBCD) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)}
\format{

}
\usage{
intercalibration_BFR
}
\description{
Intercalibration data for BFR (BDE28, BDE47, BDE99, BDE100, BDE153, BDE154,  HBCD) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)
}
\examples{
data(intercalibration_BFR)
}
\keyword{datasets}
