% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CLC.R
\docType{data}
\name{intercalibration_CLC}
\alias{intercalibration_CLC}
\title{Intercalibration data for Organochlorines (non-dioxin-like PCBs, DDTs, HCHs, HCB) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)}
\format{

}
\usage{
intercalibration_CLC
}
\description{
Intercalibration data for Organochlorines (non-dioxin-like PCBs, DDTs, HCHs, HCB) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)
}
\examples{
data(intercalibration_CLC)
}
\keyword{datasets}
