% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_metals.R
\docType{data}
\name{Intercalibration_metals}
\alias{Intercalibration_metals}
\title{Intercalibration data for metals (Hg, Ni, Cr, Pb, Cd, Zn, Cu, As, Ag) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)}
\format{

}
\usage{
Intercalibration_metals
}
\description{
Intercalibration data for metals (Hg, Ni, Cr, Pb, Cd, Zn, Cu, As, Ag) for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)
}
\examples{
data(Intercalibration_metals)
}
\keyword{datasets}
