% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_PFAS.R
\docType{data}
\name{intercalibration_PFAS}
\alias{intercalibration_PFAS}
\title{Intercalibration data for PFAS for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)}
\format{

}
\usage{
intercalibration_PFAS
}
\description{
Intercalibration data for PFAS for the Swedish national Monitoring programme for Contaminants in Marine biota (MCoM)
}
\examples{
data(intercalibration_PFAS)
}
\keyword{datasets}
