% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wide.R
\name{make_wide}
\alias{make_wide}
\title{Wide format for MCoM database.}
\usage{
make_wide(
  MCoM,
  group,
  bio = c("year", "date", "species", "species_EN", "class", "number_individuals", "sex",
    "total_length", "age", "weight", "egg_shell_thickness", "d13C", "d15N",
    "C_dw_percentage", "N_dw_percentage", "station_name", "latitude", "longitude",
    "HELCOM_basin", "ICES_basin", "season")
)
}
\arguments{
\item{MCoM}{data table of contaminants in marine biota presented in a long format.}

\item{group}{the contaminants substance groups. Choose from:  "Metals","PFAS","PCBs", "Dioxins and furans" , "Pesticides","BFRs", "Solvents-phenols","PAH". "Organotin compounds" .}

\item{bio}{choose from:"year", "date", "species", "species_EN",  "class","number_individuals","sex", "total_length","age", "weight","egg_shell_thickness" ,  "d13C", "d15N","C_dw_percentage", "N_dw_percentage", "station_name", "latitude", "longitude", "HELCOM_basin", "ICES_basin", "season"}
}
\value{
database in wide format with contaminants in separate columns.
}
\description{
Wide format for MCoM database.
}
\examples{
make_wide(MCoM, "Metals") |>  glimpse() |> View()
make_wide(MCoM, "BFRs") |> write_csv2(file = "BFRs.csv")
make_wide(MCoM, "Dioxins and furans" )

}
